source code/requirements.txt
pandas==1.0.4
python_pptx==0.6.18
pywin32==227
